requirements.txt
apache-airflow
cx_Oracle
keras
scikit-learn
pandas